+++ CriticalMapsMocksForTests/MockThemeController.swift
@@ -21,7 +21,7 @@ class MockThemeStore: ThemeStorable {
     private var currentTheme: Theme?
 
     func load() -> Theme? {
-        return currentTheme
+        currentTheme
     }
 
     func save(_ themeSelection: Theme) {