+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Changelog for Critical Maps iOS
 
+## [Unreleased]
+
+### Fixed
+
+- Fix: Don't update content when slightly swipe down in Social Modal
+- Fix whitespace only chat message can not be send anymore.
+
 ## [3.4.0] - 2019-10-08
 
 ### Added
@@ -9,17 +16,19 @@ Changelog for Critical Maps iOS
 - Landscape support
 - Set userStyle as Theme under iOS 13
 - Add infrastructure for UITests to easily generate automated screenshots with different languages and devices
+- Infrastructure for Snapshot tests
 - Open Twitter tapping on tweet
 
 ### Fixed
 
-- Fix: NavigationBar Colors under iOS 13 
+- Fix: NavigationBar Colors under iOS 13
 - Fix: UITableViewHeaderFooterView backgroundColor deprecation warning
 - Fix: Ambiguous auto layout constraints for Settings screen
 
 ## [3.3.0] - 2019-09-01
 
 ### Added
+
 - iOS 13 support
 
 ### Fixed